Fix more C warnings
There were some other warnings that were missed in the previous MR. This
one is more exhaustive, and builds cleanly for me locally. However, it
changes the public signature of the exported AVL tree functions, and as
a consequence it might break downstream dependencies. I had a look
through the other ribios* packages but I couldn’t find any that used the
AVL tree implemeentation. Still, we might want to exercise caution
before merging this change.